requirements.txt
tensorflow
scikit-learn
numpy
joblib
matplotlib
seaborn
Pillow
